refactor: move transaction tasks
