+++ client/src/eth/tx/task.rs
@@ -1,283 +1,14 @@
-use std::{cmp::max, error::Error, time::Duration};
+use std::{error::Error, time::Duration};
 
-use br_primitives::{eth::ETHEREUM_BLOCK_TIME, sub_display_format, INSUFFICIENT_FUNDS};
+use br_primitives::sub_display_format;
 use ethers::{
-	providers::{JsonRpcClient, Middleware},
+	providers::JsonRpcClient,
 	types::{TransactionReceipt, U256},
 };
 use sc_service::Arc;
 use tokio::time::sleep;
 
-use crate::eth::{
-	Eip1559GasMiddleware, EthClient, EventMessage, EventMetadata, LegacyGasMiddleware, TxRequest,
-	DEFAULT_TX_RETRIES, MAX_FEE_COEFFICIENT,
-};
-
-use super::{generate_delay, Eip1559Middleware, LegacyMiddleware};
-
-const SUB_LOG_TARGET: &str = "tx-task";
-
-/// The transaction task for Eip1559 transactions.
-pub struct Eip1559TransactionTask<T> {
-	/// The ethereum client for the connected chain.
-	pub client: Arc<EthClient<T>>,
-	/// The client signs transaction for the connected chain with local nonce manager.
-	middleware: Eip1559Middleware<T>,
-	/// The flag whether the client has enabled txpool namespace.
-	is_txpool_enabled: bool,
-	/// The minimum priority fee required.
-	min_priority_fee: U256,
-	/// If first relay transaction is stuck in mempool after waiting for this amount of time(ms),
-	/// ignore duplicate prevent logic. (default: 12s)
-	duplicate_confirm_delay: Option<u64>,
-}
-
-impl<T: JsonRpcClient> Eip1559TransactionTask<T> {
-	/// Build an Eip1559 transaction task.
-	pub fn new(
-		client: Arc<EthClient<T>>,
-		middleware: Eip1559Middleware<T>,
-		is_txpool_enabled: bool,
-		min_priority_fee: U256,
-		duplicate_confirm_delay: Option<u64>,
-	) -> Self {
-		Self { client, middleware, is_txpool_enabled, min_priority_fee, duplicate_confirm_delay }
-	}
-}
-
-#[async_trait::async_trait]
-impl<T: JsonRpcClient> AsyncTransactionTask<T> for Eip1559TransactionTask<T> {
-	fn is_txpool_enabled(&self) -> bool {
-		self.is_txpool_enabled
-	}
-
-	fn debug_mode(&self) -> bool {
-		self.client.debug_mode
-	}
-
-	fn get_client(&self) -> Arc<EthClient<T>> {
-		self.client.clone()
-	}
-
-	fn duplicate_confirm_delay(&self) -> Duration {
-		Duration::from_millis(self.duplicate_confirm_delay.unwrap_or(ETHEREUM_BLOCK_TIME * 1000))
-	}
-
-	async fn try_send_transaction(&self, mut msg: EventMessage) {
-		if msg.retries_remaining == 0 {
-			return;
-		}
-
-		// sets a random delay on external chain transactions on first try
-		if msg.give_random_delay && msg.retries_remaining == DEFAULT_TX_RETRIES {
-			sleep(Duration::from_millis(generate_delay())).await;
-		}
-
-		// set transaction `from` field
-		msg.tx_request = msg.tx_request.from(self.client.address());
-
-		// estimate the gas amount to be used
-		let estimated_gas =
-			match self.middleware.estimate_gas(&msg.tx_request.to_typed(), None).await {
-				Ok(estimated_gas) => {
-					br_metrics::increase_rpc_calls(&self.client.get_chain_name());
-					U256::from(
-						(estimated_gas.as_u64() as f64 * msg.gas_coefficient.into_f64()).ceil()
-							as u64,
-					)
-				},
-				Err(error) => {
-					return self.handle_failed_gas_estimation(SUB_LOG_TARGET, msg, &error).await
-				},
-			};
-		msg.tx_request = msg.tx_request.gas(estimated_gas);
-
-		// check the txpool for transaction duplication prevention
-		if !(self.is_duplicate_relay(&msg.tx_request, msg.check_mempool).await) {
-			// no duplication found
-			let fees = self.client.get_estimated_eip1559_fees().await;
-			let priority_fee = max(fees.1, self.min_priority_fee);
-			let max_fee_per_gas =
-				fees.0.saturating_add(priority_fee).saturating_mul(MAX_FEE_COEFFICIENT.into());
-
-			if !self.is_sufficient_funds(max_fee_per_gas, estimated_gas).await {
-				panic!(
-					"[{}]-[{}]-[{}] {}",
-					&self.client.get_chain_name(),
-					SUB_LOG_TARGET,
-					self.client.address(),
-					INSUFFICIENT_FUNDS,
-				);
-			}
-
-			let result = self
-				.middleware
-				.send_transaction(
-					msg.tx_request
-						.to_eip1559()
-						.max_fee_per_gas(max_fee_per_gas)
-						.max_priority_fee_per_gas(priority_fee),
-					None,
-				)
-				.await;
-			br_metrics::increase_rpc_calls(&self.client.get_chain_name());
-
-			match result {
-				Ok(pending_tx) => match pending_tx.await {
-					Ok(receipt) => {
-						if let Some(receipt) = receipt {
-							self.handle_success_tx_receipt(
-								SUB_LOG_TARGET,
-								Some(receipt),
-								msg.metadata,
-							);
-						} else {
-							self.handle_failed_tx_receipt(SUB_LOG_TARGET, msg).await;
-						}
-					},
-					Err(error) => {
-						self.handle_failed_tx_request(SUB_LOG_TARGET, msg, &error).await;
-					},
-				},
-				Err(error) => {
-					self.handle_failed_tx_request(SUB_LOG_TARGET, msg, &error).await;
-				},
-			}
-		}
-	}
-}
-
-/// The transaction task for Legacy transactions.
-pub struct LegacyTransactionTask<T> {
-	/// The ethereum client for the connected chain.
-	pub client: Arc<EthClient<T>>,
-	/// The client signs transaction for the connected chain with local nonce manager.
-	middleware: LegacyMiddleware<T>,
-	/// The flag whether the client has enabled txpool namespace.
-	is_txpool_enabled: bool,
-	/// The flag whether if the gas price will be initially escalated. The `escalate_percentage`
-	/// will be used on escalation. This will only have effect on legacy transactions. (default:
-	/// false)
-	is_initially_escalated: bool,
-	/// The coefficient used on transaction gas price escalation (default: 1.15)
-	gas_price_coefficient: f64,
-	/// The minimum value use for gas_price. (default: 0)
-	min_gas_price: U256,
-	/// If first relay transaction is stuck in mempool after waiting for this amount of time(ms),
-	/// ignore duplicate prevent logic. (default: 12s)
-	duplicate_confirm_delay: Option<u64>,
-}
-
-impl<T: JsonRpcClient> LegacyTransactionTask<T> {
-	/// Build an Legacy transaction task.
-	pub fn new(
-		client: Arc<EthClient<T>>,
-		middleware: LegacyMiddleware<T>,
-		is_txpool_enabled: bool,
-		is_initially_escalated: bool,
-		gas_price_coefficient: f64,
-		min_gas_price: U256,
-		duplicate_confirm_delay: Option<u64>,
-	) -> Self {
-		Self {
-			client,
-			middleware,
-			is_txpool_enabled,
-			is_initially_escalated,
-			gas_price_coefficient,
-			min_gas_price,
-			duplicate_confirm_delay,
-		}
-	}
-}
-
-#[async_trait::async_trait]
-impl<T: JsonRpcClient> AsyncTransactionTask<T> for LegacyTransactionTask<T> {
-	fn is_txpool_enabled(&self) -> bool {
-		self.is_txpool_enabled
-	}
-
-	fn debug_mode(&self) -> bool {
-		self.client.debug_mode
-	}
-
-	fn get_client(&self) -> Arc<EthClient<T>> {
-		self.client.clone()
-	}
-
-	fn duplicate_confirm_delay(&self) -> Duration {
-		Duration::from_millis(self.duplicate_confirm_delay.unwrap_or(ETHEREUM_BLOCK_TIME * 1000))
-	}
-
-	async fn try_send_transaction(&self, mut msg: EventMessage) {
-		if msg.retries_remaining == 0 {
-			return;
-		}
-
-		// sets a random delay on external chain transactions on first try
-		if msg.give_random_delay && msg.retries_remaining == DEFAULT_TX_RETRIES {
-			sleep(Duration::from_millis(generate_delay())).await;
-		}
-
-		// set transaction `from` field
-		msg.tx_request = msg.tx_request.from(self.client.address());
-
-		// estimate the gas amount to be used
-		let estimated_gas =
-			match self.middleware.estimate_gas(&msg.tx_request.to_typed(), None).await {
-				Ok(estimated_gas) => {
-					br_metrics::increase_rpc_calls(&self.client.get_chain_name());
-					U256::from(
-						(estimated_gas.as_u64() as f64 * msg.gas_coefficient.into_f64()).ceil()
-							as u64,
-					)
-				},
-				Err(error) => {
-					return self.handle_failed_gas_estimation(SUB_LOG_TARGET, msg, &error).await
-				},
-			};
-		msg.tx_request = msg.tx_request.gas(estimated_gas);
-
-		// check the txpool for transaction duplication prevention
-		if !(self.is_duplicate_relay(&msg.tx_request, msg.check_mempool).await) {
-			let mut tx = msg.tx_request.to_legacy();
-			let gas_price = self.client.get_gas_price().await;
-			if !self.is_sufficient_funds(gas_price, estimated_gas).await {
-				panic!(
-					"[{}]-[{}]-[{}] {}",
-					&self.client.get_chain_name(),
-					SUB_LOG_TARGET,
-					self.client.address(),
-					INSUFFICIENT_FUNDS,
-				);
-			}
-
-			if self.is_initially_escalated {
-				tx = tx.gas_price(
-					self.client
-						.get_gas_price_for_escalation(
-							gas_price,
-							self.gas_price_coefficient,
-							self.min_gas_price,
-						)
-						.await,
-				);
-			}
-			let result = self.middleware.send_transaction(tx, None).await;
-			br_metrics::increase_rpc_calls(&self.client.get_chain_name());
-
-			match result {
-				Ok(pending_tx) => match pending_tx.await {
-					Ok(receipt) => {
-						self.handle_success_tx_receipt(SUB_LOG_TARGET, receipt, msg.metadata)
-					},
-					Err(error) => self.handle_failed_tx_request(SUB_LOG_TARGET, msg, &error).await,
-				},
-				Err(error) => self.handle_failed_tx_request(SUB_LOG_TARGET, msg, &error).await,
-			}
-		}
-	}
-}
+use crate::eth::{EthClient, EventMessage, EventMetadata, TxRequest};
 
 #[async_trait::async_trait]
 pub trait AsyncTransactionTask<T>
